+++ Sources/MapboxDirections/RouteOptions.swift
@@ -92,6 +92,11 @@ open class RouteOptions: DirectionsOptions {
             self.maximumWeight = Measurement(value: doubleValue, unit: UnitMass.metricTons)
         }
         
+        if let mappedValue = mappedQueryItems[CodingKeys.layers.stringValue] {
+            zip(waypoints, mappedValue.components(separatedBy: ";")).forEach {
+                $0.layer = Int($1) ?? nil
+            }
+        }
         
         let formatter = DateFormatter.ISO8601DirectionsFormatter()
         if let mappedValue = mappedQueryItems[CodingKeys.departAt.stringValue],
@@ -149,6 +154,7 @@ open class RouteOptions: DirectionsOptions {
         case waypointTargets = "waypoint_targets"
         case arriveBy = "arrive_by"
         case departAt = "depart_at"
+        case layers = "layers"
     }
     
     public override func encode(to encoder: Encoder) throws {
@@ -427,6 +433,11 @@ open class RouteOptions: DirectionsOptions {
         if waypoints.first(where: { $0.targetCoordinate != nil }) != nil {
             let targetCoordinates = waypoints.filter { $0.separatesLegs }.map { $0.targetCoordinate?.requestDescription ?? "" }.joined(separator: ";")
             params.append(URLQueryItem(name: CodingKeys.waypointTargets.stringValue, value: targetCoordinates))
+        }
+        
+        if waypoints.contains(where: { $0.layer != nil }) {
+            let layers = waypoints.map { $0.layer?.description ?? "" }.joined(separator: ";")
+            params.append(URLQueryItem(name: CodingKeys.layers.stringValue, value: layers))
         }
         
         if let initialManeuverAvoidanceRadius = initialManeuverAvoidanceRadius {